refactor: mod stream_row
